Ensure google assistant code can use the intent slot=value format
