+++ lib/runner/VirtualAlexaInvoker.js
@@ -2,6 +2,7 @@ const _ = require("lodash");
 const debug = require("../util/Debug");
 const Invoker = require("./Invoker").Invoker;
 const InvokerResponse = require("./Invoker").InvokerResponse;
+const Util = require("../util/Util");
 const VirtualAlexa = require("virtual-alexa");
 
 module.exports = class VirtualAlexaInvoker extends Invoker {
@@ -104,7 +105,7 @@ module.exports = class VirtualAlexaInvoker extends Invoker {
             if (interaction.intent) {
                 response = await this._virtualAlexa.intend(interaction.intent, interaction.localizedSlots);
             } else {
-                const intent = this.detectIntent(interaction.utterance);
+                const intent = Util.returnIntentObjectFromUtteranceIfValid(interaction.utterance);
                 if (intent && this._virtualAlexa.context().interactionModel().hasIntent(intent.name)) {
                     interaction.intent = intent.name;
                     interaction.slots = intent.slots;
@@ -116,26 +117,6 @@ module.exports = class VirtualAlexaInvoker extends Invoker {
             }
         }
         return new VirtualAlexaResponse(interaction, response);
-    }
-
-    detectIntent(utter) {
-        const notEmpty = value => value && value.trim() !== "";
-        const toSlot = (accumulator, item) => {
-            const [key, val] = item.split("=");
-            accumulator[key] = val.replace(/"/g, "") ;
-            return accumulator;
-        };
-        const intentRegex = /([\w|.]*)((?: \w*=(?:"(?:[^"])*"|(?:[^ ])*))*)/;
-        const match = intentRegex.exec(utter);
-        if (match && match[0] === utter) {
-            const intent = { name: match[1] };
-            if (match[2]) {
-                const slots = match[2].split(/(\w*=(?:"(?:[^"])*"|(?:[^ ])*))/g);
-                intent.slots = slots.filter(notEmpty).reduce(toSlot, {});
-            }
-            return intent;
-        }
-        return undefined;
     }
 }
