+++ lib/runner/VirtualGoogleAssistantInvoker.js
@@ -1,6 +1,7 @@
 const _ = require("lodash");
 const Invoker = require("./Invoker").Invoker;
 const InvokerResponse = require("./Invoker").InvokerResponse;
+const Util = require("../util/Util");
 const VirtualGoogleAssistant = require("virtual-google-assistant").VirtualGoogleAssistant;
 
 module.exports = class VirtualGoogleAssistantInvoker extends Invoker {
@@ -57,6 +58,15 @@ module.exports = class VirtualGoogleAssistantInvoker extends Invoker {
             if (interaction.intent) {
                 response = await this._virtualGoogleAssistant.intend(interaction.intent, interaction.slots);
             } else {
+                const intent = Util.returnIntentObjectFromUtteranceIfValid(interaction.utterance);
+
+                if (intent) {
+                    interaction.intent = intent.name;
+                    interaction.slots = intent.slots;
+                    response = await this._virtualGoogleAssistant.intend(interaction.intent, interaction.slots);
+                } else {
+                    response = await this._virtualGoogleAssistant.utter(interaction.utterance)
+                }
                 response = await this._virtualGoogleAssistant.utter(interaction.utterance)
             }
         }
