+++ lib/util/Util.js
@@ -91,6 +91,26 @@ module.exports = class Util {
 
     }
 
+    static returnIntentObjectFromUtteranceIfValid(utterance) {
+        const notEmpty = value => value && value.trim() !== "";
+        const toSlot = (accumulator, item) => {
+            const [key, val] = item.split("=");
+            accumulator[key] = val.replace(/"/g, "") ;
+            return accumulator;
+        };
+        const intentRegex = /([\w|.]*)((?: \w*=(?:"(?:[^"])*"|(?:[^ ])*))*)/;
+        const match = intentRegex.exec(utterance);
+        if (match && match[0] === utterance) {
+            const intent = { name: match[1] };
+            if (match[2]) {
+                const slots = match[2].split(/(\w*=(?:"(?:[^"])*"|(?:[^ ])*))/g);
+                intent.slots = slots.filter(notEmpty).reduce(toSlot, {});
+            }
+            return intent;
+        }
+        return undefined;
+    }
+
     static formatDate(date) {
          return date.getFullYear() + "-" + Util.padZero(date.getMonth()+1)+ "-" + Util.padZero(date.getDate()) + "T" +
              Util.padZero(date.getHours()) + ":" + Util.padZero(date.getMinutes()) + ":" +
